requirements.txt
Babel==2.3.4
Django==1.9.7
django-authtools==1.5.0
django-braces==1.9.0
django-phonenumber-field==1.1.0
phonenumberslite==7.4.5
psycopg2==2.6.1
pytz==2016.6.1
Unipath==1.1
